Extract source visualizer logic to adjust `ThreadFront` import
