+++ src/devtools/client/debugger/src/components/Editor/SourcemapToggle.tsx
@@ -10,7 +10,7 @@ import classNames from "classnames";
 import { setModal } from "ui/actions/app";
 import { ThreadFront } from "protocol/thread";
 import { SourceId } from "@recordreplay/protocol";
-import { getUniqueAlternateSourceId } from "../../utils/source";
+import { getUniqueAlternateSourceId } from "../../utils/sourceVisualizations";
 
 function SourcemapError({ onClick }: { onClick: () => void }) {
   return (
