+++ src/devtools/client/debugger/src/utils/source.d.ts
@@ -2,15 +2,6 @@ import { SourceId } from "@recordreplay/protocol";
 import { Source, SourceContent, Location } from "../reducers/sources";
 import { AsyncValue } from "./async-value";
 
-export function getSourcemapVisualizerURL(
-  selectedSource: Source,
-  alternateSource: Source
-): string | null;
-export function getUniqueAlternateSourceId(sourceId: SourceId): {
-  sourceId?: SourceId;
-  why?: "no-sourcemap" | "not-unique";
-};
-
 export function getTextAtPosition(
   asyncContent: SourceContent | AsyncValue<SourceContent>,
   location: Location
@@ -23,6 +14,9 @@ export function getPlainUrl(url: string): string;
 export function getRawSourceURL(url: string): string;
 
 export function isInlineScript(source: { introductionType?: string | null }): boolean;
+
+export function isNodeModule(source: Source): boolean;
+export function isBowerComponent(source: Source): boolean;
 
 export function getFilename(source: Source, rawSourceURL?: string): string;
